% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\docType{data}
\name{people}
\alias{people}
\title{people: A registry of all people that have worked for Global Health Engineering, including undergraduate student projects.}
\format{
A tibble with 129 rows and 7 variables
\describe{
\item{title}{ Unique identifier for each person using a combination of other metadata. Folder name on Google Drive.}
\item{degree}{ Categorical variable with four levels: bsc, msc, phd, staff.}
\item{type}{ Categorical variable with six levels: hiwi, intern, post-doc, scientific-assitant, sem-proj, thesis.}
\item{b_m_student}{ Binary variable to identify if person is BSc or MSc students. Levels: yes, no.}
\item{start_date}{ Start date of the person.}
\item{year}{ Year of the start date.}
\item{thesis_title}{ Title of the thesis.}
}
}
\usage{
people
}
\description{
This dataset contains the following columns:
}
\keyword{datasets}
